- Added new queries - moved all the transformer into the GraphQLTransformers file as package level functions
